refactor : 넘버 맵수정
- OrderProcess에서 Pos로 이동
