+++ src/main/java/domain/PayProcess.java
@@ -1,7 +1,27 @@
 package domain;
 
+import java.util.HashMap;
+import java.util.List;
+
 public class PayProcess {
-  public void start(){
+
+  private final List<Table> tables;
+  private final List<Menu> menus;
+
+  private final HashMap<Integer, Integer> tableNumbers;
+  private final HashMap<Integer, Integer> menuNumbers;
+
+  public PayProcess(List<Table> tables, List<Menu> menus, HashMap<Integer, Integer> tableNumbers,
+      HashMap<Integer, Integer> menuNumbers) {
+    this.tables = tables;
+    this.menus = menus;
+    this.tableNumbers = tableNumbers;
+    this.menuNumbers = menuNumbers;
+
+
+  }
+
+  public void start() {
 
   }
 }
